+++ src/main/java/project/unsw/gloriaromanus/GloriaRomanusLauncher.java
@@ -1,0 +1,8 @@
+package project.unsw.gloriaromanus;
+
+public class GloriaRomanusLauncher {
+
+    public static void main(String[] args) {
+        GloriaRomanusApplication.main(args);
+    }
+}
