+++ src/main/java/project/unsw/gloriaromanus/initial_province_ownership.json
@@ -1,0 +1,59 @@
+{
+  "Rome": [
+    "Lugdunensis",
+    "Lusitania",
+    "Lycia et Pamphylia",
+    "Macedonia",
+    "Mauretania Caesariensis",
+    "Mauretania Tingitana",
+    "Moesia Inferior",
+    "Moesia Superior",
+    "Narbonensis",
+    "Noricum",
+    "Numidia",
+    "Pannonia Inferior",
+    "Pannonia Superior",
+    "Raetia",
+    "Sardinia et Corsica",
+    "Sicilia",
+    "Syria",
+    "Tarraconensis",
+    "Thracia",
+    "V",
+    "VI",
+    "VII",
+    "VIII",
+    "X",
+    "XI"
+  ],
+  "Gaul": [
+    "Achaia",
+    "Aegyptus",
+    "Africa Proconsularis",
+    "Alpes Cottiae",
+    "Alpes Graiae et Poeninae",
+    "Alpes Maritimae",
+    "Aquitania",
+    "Arabia",
+    "Armenia Mesopotamia",
+    "Asia",
+    "Baetica",
+    "Belgica",
+    "Bithynia et Pontus",
+    "Britannia",
+    "Cilicia",
+    "Creta et Cyrene",
+    "Cyprus",
+    "Dacia",
+    "Dalmatia",
+    "Galatia et Cappadocia",
+    "Germania Inferior",
+    "Germania Superior",
+    "I",
+    "II",
+    "III",
+    "IV",
+    "IX",
+    "Iudaea"
+  ]
+}
