--- src/unsw/gloriaromanus/GloriaRomanusLauncher.java
@@ -1,8 +1,0 @@
-package unsw.gloriaromanus;
-
-public class GloriaRomanusLauncher {
-
-    public static void main(String[] args) {
-        GloriaRomanusApplication.main(args);
-    }
-}
